Add files via upload
In today's hectic environment, recommendation systems are becoming more and more crucial. People never have enough time because of all the things they have to get done in a 24-hour period. Recommendation systems are crucial because they save users' cognitive resources by assisting them in making wise decisions.
A recommendation system's main goal is to find material that a certain user could find interesting. Moreover, it takes a variety of elements to provide each user or individual with personalized lists of relevant and engaging material. Artificial intelligence-driven algorithms, known as recommendation systems, go through all of the available possibilities to provide a personalised list of goods that are both interesting and pertinent to a certain person. Based on their profile, past searches and browsing activity, what other individuals with comparable characteristics and demographics are viewing, and your propensity to watch those films, these results are shown. Using the existing data, heuristics and predictive modelling are used to do this.